Allow smaller min size to work for chisel & bits
I know you previously stated that you don't wish to allows players to x-ray so you set a min size, but shouldn't this just be left up to the player themselves or the server owner? Personally, I'd enjoy this mod much more if I could shrink to a smaller size. Is there a way I could edit the code myself to do this personally without you having to change the entire mod? Thanks, in advance.